Task 8 - Employee Promotion Evaluation.
Write a recursive function to calculate years to promotion to reach level 10. First it will establish what level they are at and if they are at 10 or above it will return 0, if they are below then it will calculate the years they have worked and how many years left until level 10.