Task 8 - Employee Promotion Evaluation (UPDATED VERSION)
The recursive functions for task 8 was edited so that the function can be called back in itself to properly calculate the current number of years the employee has worked based on the employee level plugged in and how many years until they reach level 10. The function will run through itself until the employee level 10 is reached while adding 2 years (the number of years per level) each time a level is added.